backend/requirements.txt
websockets
ollama
psutil
requests
beautifulsoup4
pytesseract
Pillow
PyYAML
